+++ TelegramBots.Common/Services/TelegramBotService.cs
@@ -13,9 +13,9 @@ namespace TelegramBots.Common.Services
     public interface ITelegramBotService<T> where T: TelegramBotInstance
     {
         Task<Update> GetUpdate(int lastUpdateId);
-        Task<Message> SendTextMessage(int userId, string text, ParseMode parseMode = ParseMode.Html);
-        Task<Message> SendInlineButtonMessage(int userId, string text, InlineKeyboardMarkup markup);
-        Task SendAudioMessage(int userId, string audioLink, string name);
+        Task<Message> SendTextMessage(long userId, string text, ParseMode parseMode = ParseMode.Html);
+        Task<Message> SendInlineButtonMessage(long userId, string text, InlineKeyboardMarkup markup);
+        Task SendAudioMessage(long userId, string audioLink, string name);
         Task SetWebHook(string url);
     }
 
@@ -38,12 +38,12 @@ namespace TelegramBots.Common.Services
                 .FirstOrDefault();
         }
 
-        public async Task<Message> SendTextMessage(int userId, string text, ParseMode parseMode = ParseMode.Html)
+        public async Task<Message> SendTextMessage(long userId, string text, ParseMode parseMode = ParseMode.Html)
         {
             return await _bot.SendTextMessageAsync(new ChatId(userId), text, parseMode);
         }
 
-        public async Task SendAudioMessage(int userId, string audioLink, string name)
+        public async Task SendAudioMessage(long userId, string audioLink, string name)
         {
             try
             {
@@ -57,7 +57,7 @@ namespace TelegramBots.Common.Services
             }
         }
 
-        public async Task<Message> SendInlineButtonMessage(int userId, string text, InlineKeyboardMarkup markup)
+        public async Task<Message> SendInlineButtonMessage(long userId, string text, InlineKeyboardMarkup markup)
         {
             return await _bot.SendTextMessageAsync(new ChatId(userId), text,
                 replyMarkup: markup, parseMode: ParseMode.Html);
